refactor(TestBed): extract platform and view-container overrides
